feat: add solution for task Modules
